+++ buildSrc/src/main/kotlin/dev/muon/medieval/gradle/Versions.kt
@@ -1,7 +1,7 @@
 package dev.muon.medieval.gradle
 
 object Versions {
-    const val MOD = "1.8.0"
+    const val MOD = "1.8.1"
 
     const val MINECRAFT = "1.21.1"
     const val PARCHMENT_MINECRAFT = "1.21"
@@ -39,7 +39,7 @@ object Versions {
     const val APOTHIC_ENCHANTING = "1.2.5"
     const val PLACEBO = "9.6.2"
 
-    const val SPELL_ENGINE = "1.3.2"
-    const val SPELL_POWER = "1.1.2"
+    const val SPELL_ENGINE = "1.5.1"
+    const val SPELL_POWER = "1.2.1"
     const val TINY_CONFIG = "2.3.2"
 }